props type updated
