+++ src/components/CustomCheckBox.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { View, Text, TouchableOpacity, StyleSheet } from 'react-native';
+import { View, TouchableOpacity, StyleSheet } from 'react-native';
 
 interface CustomCheckBoxProps {
   checked: boolean;
@@ -11,7 +11,7 @@ const Checkbox: React.FC<CustomCheckBoxProps> = ({ checked, onPress }) => {
     <TouchableOpacity
       accessibilityRole="button"
       onPress={onPress}
-      className="border-cyan-30 my-auto border bg-transparent p-0.5 rounded"
+      className="border-cyan-30 my-auto rounded border bg-transparent p-0.5"
     >
       {checked ? (
         <View style={styles.checkedContainer} />
@@ -27,7 +27,7 @@ const styles = StyleSheet.create({
     width: 10,
     height: 10,
     backgroundColor: 'black',
-    borderRadius:2,
+    borderRadius: 2,
   },
   uncheckedContainer: {
     width: 10,
