+++ src/components/ExerciseCard.tsx
@@ -2,24 +2,27 @@
 import React from 'react';
 import { View, Text, StyleSheet, ScrollView } from 'react-native';
 import { useFonts, BebasNeue_400Regular } from '@expo-google-fonts/bebas-neue';
-import NoteBlock from './ExerciseNoteBlock';
-import RestBlock from './ExerciseRestBlock';
-import SetBlock from './ExerciseSetBlock';
-
+import NoteBlock from './blocks/ExerciseNoteBlock';
+import RestBlock from './blocks/ExerciseRestBlock';
+import SetBlock from './blocks/ExerciseSetBlock';
+import { Exercise, ExerciseItem, Set, Note, Rest } from '../../types';
+import {isSet, isNote, isRest} from '../utils/typeCheck';
 {/**WILL CHANGE PROPS*/}
-interface ExerciseCardProps {
-    exerciseProp: Array<any>;
+export interface ExerciseCardProps {
+    exercise: Exercise;
 }
 
 
 {/**WILL CHANGE PROPS*/}
-const ExerciseCard:React.FC<ExerciseCardProps> = ({exerciseProp}) => {
-    const [currentSetsDone, setCurrentSetDone] = React.useState<number>();
-    const [maxSets, setMaxSets] = React.useState<number>();
+const ExerciseCard:React.FC<ExerciseCardProps> = ({exercise}) => {
+    const maxSets = React.useRef(calculateNumWorkingSets(exercise.items));
+    const [currentSetsDone, setCurrentSetDone] = React.useState<number>(0);
+    
     const [userInputReps, setUserInputReps] = React.useState<string>('');
     const [userInputWeight, setUserInputWeight] = React.useState<string>('');
+
+    React.useEffect(() => {maxSets.current = calculateNumWorkingSets(exercise.items);},[exercise]);
     
-    React.useState(() => {setCurrentSetDone},[userInputReps, userInputWeight])
     // Load font
     const [fontsLoaded] = useFonts({
         BebasNeue_400Regular,
@@ -33,16 +36,25 @@ const ExerciseCard:React.FC<ExerciseCardProps> = ({exerciseProp}) => {
       <View style={styling.container} className='mt-5 h-5/6 w-11/12 bg-white shadow-sm rounded-xl'>
         <ScrollView>
             <View className='h-10 w-full mb-2 content-evenly border-b border-slate-200 flex-row justify-between'>
-                <Text className="m-1 mt-3 ml-3 font-bold text-center"> Deadlift</Text>
-                <Text style={currentSetsDone==maxSets?styling.greenText:styling.redText} className="m-1 mt-3 mr-3 font-bold"> {currentSetsDone} / {maxSets} Sets Done</Text>
+                <Text className="m-1 mt-3 ml-3 font-bold text-center">{exercise.name}</Text>
+                <Text style={currentSetsDone==maxSets.current?styling.greenText:styling.redText} className="m-1 mt-3 mr-3 font-bold"> {currentSetsDone} / {maxSets.current} Sets Done</Text>
             </View>
             {/** TEST CODE, WILL REMOVE LATER */}
-            <NoteBlock note="asd"/>
-            <RestBlock minutes="1" seconds='30'/>
-            <SetBlock setNumber={1} reps={'5-9'} rpe="10" orm="100%" warmup={false} getUserRecordedValue={(reps,weight) => {
-                setUserInputReps(reps);
-                setUserInputWeight(weight);
-            }}/>
+            {exercise.items.map((item, index) => { 
+                if(isSet(item)) {
+                    return(<SetBlock setNumber={index} reps={item.reps} rpe={item.rpe} orm={item.orm} warmup={item.warmup}/>);
+                }
+                else if(isRest(item)) {
+                    return(<RestBlock minutes={item.minutes} seconds={item.seconds}/>);
+                }
+                else if(isNote(item)) {
+                    return(<NoteBlock note={item.text}/>);
+                } else {
+                    return(<View className='flex flex-grow border h-10'>
+                            <Text className="font-bold text-red-700 text-center my-auto">**Item Failed to Load**</Text>
+                        </View>);
+                }
+            })}
             {/** END OF TEST CODE, WILL REMOVE LATER */}
         </ScrollView>
       </View>
@@ -60,5 +72,17 @@ const styling = StyleSheet.create({
         color: '#ED4E39',
     },
 }); 
+
+function calculateNumWorkingSets(arr: ExerciseItem[]): number {
+    let numWorkingSets = 0;
+    for(let i = 0; i<arr.length; i++) {
+        if (isSet(arr[i])) {
+            if(!arr[i].warmup)  {
+                numWorkingSets++;
+            }
+        }
+    }    
+    return numWorkingSets;
+}
 
 export default ExerciseCard;
